folder structure changed left section in product list page completed
