+++ src/Pages/Product_List_Page/Product_List_Right/Product_List_Right.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import "./Product_List_Right.css"
+
+const Product_List_Right = () => {
+  return (
+    <>
+      <div>Product_List_Right</div>
+    </>
+  );
+}
+
+export default Product_List_Right
